topdown: Ignore errors during parsing Expires header
Earlier if the value for the Expires header was set
to an invalid date like 0, we would return an error and
fallback to the normal cache (ie.we would not use
the inter-query cache).

Servers can sometimes set `Expires: 0` to indicate
expired content. This change ignores the errors
resulting from parsing the Expires header.

Fixes: #3284

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>